+++ packages/flow-core/CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ### Improvements
 
-- - Added `category` property to `f-tag` to support both `fill` and `outline` values. Default value set to `fill`.
+- Added `category` property to `f-tag` to support both `fill` and `outline` values. Default value set to `fill`.
+- Added `max-width` property to `f-tag`.
 
 ## [2.9.6] - 2024-03-28